requirements.txt
# Advanced Trading System - Függőségek
# Raspberry Pi 4 optimalizált

# Alap csomagok
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-RESTful==0.3.10
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1

# Adatbázis
SQLAlchemy==2.0.27
alembic==1.13.1
pymysql==1.1.0
sqlite3-api==0.1.0

# Adatfeldolgozás
numpy==2.2.6
pandas==2.2.3
scipy==1.12.0
scikit-learn==1.4.0
statsmodels==0.14.1
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.1.2
dash==2.16.1

# Kriptovaluta API-k
ccxt==4.2.15
python-binance==1.0.19
krakenex==2.1.0
coinbase==2.1.0

# Gépi tanulás
tensorflow-lite==2.15.0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.3
nltk==3.8.1
textblob==0.17.1
transformers==4.38.2

# Hálózat és API
requests==2.32.3
aiohttp==3.9.3
websockets==12.0
urllib3==2.4.0
beautifulsoup4==4.13.4

# Segédeszközök
python-dotenv==1.0.1
pyyaml==6.0.2
schedule==1.3.0
apscheduler==3.10.4
tqdm==4.67.1
joblib==1.3.2
pytz==2025.2
python-dateutil==2.9.0.post0
psutil==5.9.8
cryptography==45.0.3

# Értesítések
python-telegram-bot==20.8
discord.py==2.3.2
sendgrid==6.11.0
twilio==8.12.0

# Tesztelés
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Dokumentáció
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3

# Optimalizálás (RPI4 specifikus)
numba==0.59.0
dask==2023.12.1
memory-profiler==0.61.0
line-profiler==4.1.1

# Web fejlesztés
gunicorn==21.2.0
eventlet==0.35.0
gevent==23.9.1
